fix(route): fix gov/sichuan/deyang/govpulicinfo
