+++ lib/v2/gov/router.js
@@ -61,7 +61,7 @@ module.exports = function (router) {
     router.get('/shenzhen/hrss/szksy/:caty/:page?', require('./shenzhen/hrss/szksy/index'));
     router.get('/shenzhen/xxgk/zfxxgj/:caty', require('./shenzhen/xxgk/zfxxgj'));
     router.get('/shenzhen/zzb/:caty/:page?', require('./shenzhen/zzb/index'));
-    router.get('/sichuan/deyang/govpulicinfo/:countyName', require('./sichuan/deyang/govpulicinfo'));
+    router.get('/sichuan/deyang/govpulicinfo/:countyName/:infoType?', require('./sichuan/deyang/govpublicinfo'));
     router.get('/taiyuan/rsj/:caty/:page?', require('./taiyuan/rsj'));
     router.get('/wuhan/sy/whyw', require('./wuhan/whyw'));
     router.get(/xinyi(\/[\w/-]+)?/, require('./xinyi/xinyi'));
